update: music search
